Break trace module into smaller parts.
